+++ packages/tools/tests/test/audioV2/staticSound.playback.test.ts
@@ -1,5 +1,5 @@
 import { AddSharedAbstractSoundPlaybackTests } from "./shared/abstractSound.playback";
-import { GetPulseCounts, InitAudioEngineV2Tests, L } from "./utils/audioEngineV2.utils";
+import { GetPulseCounts, InitAudioEngineV2Tests, L } from "./utils/audioV2.utils";
 
 import { expect, test } from "@playwright/test";
