+++ packages/tools/tests/test/audioV2/streamingSound.playback.test.ts
@@ -1,5 +1,5 @@
 import { AddSharedAbstractSoundPlaybackTests } from "./shared/abstractSound.playback";
-import { InitAudioEngineV2Tests } from "./utils/audioEngineV2.utils";
+import { InitAudioEngineV2Tests } from "./utils/audioV2.utils";
 
 InitAudioEngineV2Tests();
 AddSharedAbstractSoundPlaybackTests("Streaming");
